kubeconfig: move to new package: cmd/wksctl/kubeconfig
