+++ cmd/wksctl/main.go
@@ -10,6 +10,7 @@ import (
 	"github.com/weaveworks/wksctl/cmd/wksctl/addon"
 	"github.com/weaveworks/wksctl/cmd/wksctl/apply"
 	"github.com/weaveworks/wksctl/cmd/wksctl/bashcompletions"
+	"github.com/weaveworks/wksctl/cmd/wksctl/kubeconfig"
 	"github.com/weaveworks/wksctl/cmd/wksctl/plan"
 	"github.com/weaveworks/wksctl/cmd/wksctl/registrysynccommands"
 	"github.com/weaveworks/wksctl/cmd/wksctl/version"
@@ -42,6 +43,7 @@ func main() {
 
 	rootCmd.AddCommand(addon.Cmd)
 	rootCmd.AddCommand(apply.Cmd)
+	rootCmd.AddCommand(kubeconfig.Cmd)
 	rootCmd.AddCommand(plan.Cmd)
 	rootCmd.AddCommand(registrysynccommands.Cmd)
 	rootCmd.AddCommand(version.Cmd)
